api/src/pyproject.toml
[tool.poetry]
name = "balhom-drf-api"
version = "2.0.0"
description = ""
authors = ["Fabbo"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-stubs = "^4.2.3"
django-configurations = "^2.4.1"
django-filter = "^23.2"
django-cors-headers = "^4.2.0"
drf-yasg = "^1.21.7"
pyyaml = "^6.0.1"
coreapi = "^2.3.3"
python-keycloak = "^3.3.0"
pyjwt = "^2.8.0"
pillow = "^10.0.0"
celery = "^5.3.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
redis = "^4.6.0"
eventlet = "^0.33.3"
uritemplate = "^4.1.1"
psycopg2-binary = "^2.9.7"
coverage = "^7.2.7"
beautifulsoup4 = "^4.12.2"
gunicorn = "^21.2.0"
minio = "^7.1.15"
django-storages = "^1.13.2"
boto3 = "^1.28.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
